ENH: add initial version of external sort using k-way merge

+++ src/lib/record.h
@@ -3,8 +3,8 @@
 #include <array>
 #include <cstdint>
 
-const int TUPLE_SIZE = 100;
-const int KEY_SIZE = 10;
+const size_t TUPLE_SIZE = 100;
+const size_t KEY_SIZE = 10;
 
 using Record = std::array<uint8_t, TUPLE_SIZE>;
 using Header = std::array<uint8_t, KEY_SIZE>;
